Create wssclient.js
This client establishes a connection to web socket server which has connection to devices such as peanut plug etc. The client waits for voice-based command from server (in this case, the localhost server) and on receipt of the command, sends a message to web socket server with the details of the device and on/off switch. It waits for the feedback from the web socket server about the status of the device and in turn sends the feedback details to the local host server.